Convert to LangVersion 12 as max supported for .NET 8.0
Many of these projects used `<LangVersion13</LangVersion>` However, that language version is greater than what is supported by the max runtime they are configured to target. This PR converts them to using at most LangVersion 12, supported by .NET8 LTS runtime. (#11)

* Ubiquity.NET.CodeAnaysis.Utils is the exception as that is exclusively targeting RoslynExtensions and only supports .NET standard 2.0.
    - It is still set to a max of 12 and the rest is poly filled.

Most of the changes surround use of params collections instead of just arrays. In all such cases an `#if` is used to detect the support for that feature and falls back to a method/constructor using an array that forwards (tail calls) an Enumerator version) IFF the runtime supports that feature is it used. This should have no impact on consumers of this library except that it doesn't lie and won't use features not supported by the language/runtime.