Added .NET 10 SDK use and targeting support. * Fixes #15 * Updated DocFx to use built binaries     - Bugs in current DocFx don't allow generating from sources using newer SDK.     - Sadly other bugs in DocFx produce warnings for the binary builds...     - Ultimately, at the very least, the .NET metadata generation needs excising from DocFx to a standalone tool that is incrementally versioned. This would allow for more rapid response to changes in language and runtime. * Doc comment corrections/clarifications * Added .NET10 targeting for some projects * Removed redundant internal ReportingTextWriter type     - Fixes #13 * Excluded additional tests from code coverage.     - Ideally a simple one-time assembly attribute is enough, but sadly that doesn't exist for .NET standard 2.0 so each test class must add that * Removed use of `LangVersion` property for all but projects targeting .NET Standard 2.0.     - Those are using LangVersion 12, which is the max supported by .NET 8
Sadly, this does increase code maintenance complexity as the multiple layers of `#if/#else/#endif` make it much more difficult.

Projects that have no benefits from either the `extension` or `field` keywords stay on .NET 8 as the max runtime. The `field` keyword does not require runtime support and is entirely "internal'. However the `extension` keyword is NOT unless the extensions are internal. For a public extension the consumer must also support that so use is limited to ONLY when targeting .NET 10